FORWORK fix? ajustes finais

+++ src/main/java/com/thiago/forwork/dto/user/UserSaveDTO.java
@@ -1,5 +1,5 @@
 package com.thiago.forwork.dto.user;
 
 public record UserSaveDTO(
-    String name, Long department
+    String name, Integer department
 ){}
